feat(frontend): Duplicate component
