+++ frontend/src/lib/components/apps/utils.ts
@@ -2,11 +2,19 @@ import type { Schema } from '$lib/common'
 import { FlowService, ScriptService } from '$lib/gen'
 import { inferArgs } from '$lib/infer'
 import { emptySchema } from '$lib/utils'
-import type { AppComponent } from './editor/component'
+import type { AppComponent, AppComponentConfig } from './editor/component'
+
+import {
+	components as componentsRecord,
+	getRecommendedDimensionsByComponent
+} from './editor/component'
+import { gridColumns } from './gridUtils'
+import gridHelp from '@windmill-labs/svelte-grid/src/utils/helper'
 
 import type { AppInput, InputType, ResultAppInput, StaticAppInput } from './inputType'
 import type { Output } from './rx'
 import type { App, GridItem } from './types'
+import { getNextId } from '../flows/flowStateUtils'
 
 export async function loadSchema(
 	workspace: string,
@@ -64,8 +72,6 @@ export function schemaToInputsSpec(
 		return accu
 	}, {})
 }
-
-
 
 export function accessPropertyByPath<T>(object: T, path: string): T | undefined {
 	// convert indexes to properties
@@ -202,13 +208,111 @@ export function getAllScriptNames(app: App): string[] {
 }
 
 function clearAndUpper(text: string) {
-	return text.replace(/-/, "").toUpperCase();
+	return text.replace(/-/, '').toUpperCase()
 }
 
 export function toPascalCase(text: string) {
-	return text.replace(/(^\w|-\w)/g, clearAndUpper);
+	return text.replace(/(^\w|-\w)/g, clearAndUpper)
 }
 
 export function toKebabCase(text: string) {
-	return text.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
+	return text.replace(/([a-z0-9])([A-Z])/g, '$1-$2').toLowerCase()
+}
+
+export function findParent(root: GridItem[], id: string): GridItem | undefined {
+	for (const a of root) {
+		if (a.id === id) {
+			return a
+		}
+
+		if (a.data.subGrids) {
+			// Recursively search the sub-grids
+			for (const subGrid of a.data.subGrids) {
+				const result = findParent(subGrid, id)
+				if (result) {
+					return result
+				}
+			}
+		}
+	}
+
+	return undefined
+}
+
+export function insertNewGridItem(
+	root: GridItem[],
+	id: string,
+	subGridIndex: number,
+	newId: string,
+	data
+): GridItem[] {
+	const parentA = findParent(root, id)
+
+	if (!parentA) {
+		throw new Error(`Parent A object with ID ${id} not found.`)
+	}
+
+	const subGrid = parentA.data.subGrids[subGridIndex]
+
+	if (!subGrid) {
+		throw new Error(`Sub-grid with index ${subGridIndex} not found.`)
+	}
+
+	const newItem = createNewGridItem(subGrid ?? [], newId, data)
+	subGrid.push(newItem)
+	return root
+}
+
+// The grid is needed to find a space for the new component
+export function createNewGridItem(grid: GridItem[], id: string, data): GridItem {
+	const appComponent = data
+
+	appComponent.id = id
+
+	const newComponent = {
+		fixed: false,
+		resizable: true,
+		draggable: true,
+		customDragger: false,
+		customResizer: false,
+		x: 0,
+		y: 0
+	}
+
+	let newData: AppComponent = JSON.parse(JSON.stringify(appComponent))
+
+	const newItem: GridItem = {
+		data: newData,
+		id: id
+	}
+
+	gridColumns.forEach((column) => {
+		const rec = getRecommendedDimensionsByComponent(appComponent.type, column)
+
+		newItem[column] = {
+			...newComponent,
+			min: { w: 1, h: 1 },
+			max: { w: column, h: 100 },
+			w: rec.w,
+			h: rec.h
+		}
+		const position = gridHelp.findSpace(newItem, grid, column) as { x: number; y: number }
+		newItem[column] = { ...newItem[column], ...position }
+	})
+
+	return newItem
+}
+
+function recursiveGetIds(gridItem: GridItem): string[] {
+	const subGrids = gridItem.data.subGrids ?? []
+	const subGridIds = subGrids
+		.map((subGrid: GridItem[]) => subGrid.map(recursiveGetIds))
+		.flat(Infinity)
+	return [gridItem.data.id, ...subGridIds]
+}
+
+export function getNextGridItemId(grid: GridItem[] = []): string {
+	const gridItemIds = grid.map(recursiveGetIds).flat()
+	const id = getNextId(gridItemIds)
+	return id
 }
